+++ functions/changeStretch.ts
