--- index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-
-
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + Vue + TS</title>
-  </head>
-
-
-  <body>
-    <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
-    <script
-    language="JavaScript"
-    type="text/javaScript"
-    src="./lib/fits.js"
-  ></script>
-  </body>
-</html>
