+++ package.json
@@ -1,25 +1,17 @@
 {
-  "name": "clearsky",
   "private": true,
-  "version": "0.0.0",
-  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "preview": "vite preview --port 8080"
-  },
-  "dependencies": {
-    "@heroicons/vue": "^2.0.1",
-    "pinia": "^2.0.17",
-    "vue": "^3.2.37"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^3.0.0",
-    "autoprefixer": "^10.4.8",
-    "postcss": "^8.4.16",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.6.4",
-    "vite": "^3.0.0",
-    "vue-tsc": "^0.38.4"
+    "@unocss/nuxt": "^0.45.21",
+    "nuxt": "3.0.0-rc.9"
+  },
+  "dependencies": {
+    "@pinia/nuxt": "^0.4.2"
   }
 }
