+++ public/index.css
@@ -1,0 +1,27 @@
+:root {
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+    font-size: 16px;
+    color: rgba(255, 255, 255, 0.87);
+    background-color: #000000;
+}
+
+.header {
+    width: 100px;
+    height: 60px;
+    font-size: 18px;
+}
+
+.export {
+    padding-top: 10px;
+    padding-bottom: 10px;
+    padding-left: 12px;
+    padding-right: 12px;
+    border-radius: 10px;
+    height: 60px;
+    background-color: white;
+    color: #000000;
+}
+
+#stretch {
+    background-color: #100f0f;
+}
