--- readme.es.md
@@ -1,27 +1,0 @@
-# cosmoFits 🪐
-
-## El proyecto cosmoFits permite la lectura y modificación de los archivos .fts
-
-### Idioma:
-- 🇬🇧 [English](./README.md)
-- 🇲🇽 Español
-
-## Instalación
-
-### Clona el proyecto:
-
-```
-git clone https://github.com/codingTeach/cosmoFit
-```
-
-### Dirigete al directorio del proyecto e instala las dependencias:
-
-```
-cd cosmoFit && npm i
-```
-
-### Ejecuta cosmoFits(desarrollo) 🔥
-
-```
-npm run dev
-```
