--- src/main.ts
@@ -1,6 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import { createPinia } from 'pinia'
-import './index.css'
-
-createApp(App).use(createPinia()).mount('#app')
