--- tailwind.config.cjs
@@ -1,13 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-
-    content: [
-      "./index.html",
-      "./src/**/*.{vue,js,ts,jsx,tsx}",
-    ],
-
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
