--- vite.config.ts
@@ -1,10 +1,0 @@
-import { build, defineConfig } from 'vite'
-import vue from '@vitejs/plugin-vue'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [vue()],
-  
-  //assetsInclude: './lib/fits.js'
-  
-})
